feat: Implementa task para consulta de informações de PIX perante o Bacen
